+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Breaking Changes
 
 ### Bugs Fixed
+* Rpc tags linked to `OperationType.Rpc`
+  ([#44906](https://github.com/Azure/azure-sdk-for-net/pull/44906))
 
 ### Other Changes